Add logging.FileHandler only if writeable
When installing PRSedm as a root-owned package or in a container,
writing to the log file may not be possible by a non-superuser.

We add a check to see if the path is writeable before adding
logging.FileHandler(log_path) to the list of log handlers.

If not writeable, the logging.StreamHandler() will simply log
to stderr.